+++ src/lib/components/DividerWithArrow.tsx
@@ -4,8 +4,8 @@ import { CustomIcon } from "lib/components/icon";
 
 export const DividerWithArrow = () => (
   <Flex gap={2} alignItems="center" py={3}>
-    <Divider borderColor="accent.main" />
-    <CustomIcon name="arrow-down" boxSize={4} color="accent.main" />
-    <Divider borderColor="accent.main" />
+    <Divider borderColor="primary.main" />
+    <CustomIcon name="arrow-down" boxSize={4} color="primary.main" />
+    <Divider borderColor="primary.main" />
   </Flex>
 );