+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/messages/swap/MsgSwapExactAmountOutAction.tsx
@@ -36,7 +36,7 @@ export const MsgSwapExactAmountOutAction = ({
         fontWeight={400}
         ampCopierSection={ampCopierSection}
       />
-      <CustomIcon name="arrow-right" boxSize={4} color="accent.main" />
+      <CustomIcon name="arrow-right" boxSize={4} color="primary.main" />
       <MsgToken
         token={outToken}
         fontWeight={700}