+++ src/lib/styles/theme/components/textarea.ts
@@ -14,11 +14,11 @@ export const Textarea: ComponentStyleConfig = {
         borderColor: "gray.700",
       },
       _focusWithin: {
-        borderColor: "secondary.main",
+        borderColor: "primary.dark",
         borderWidth: "2px",
       },
       _active: {
-        borderColor: "secondary.main",
+        borderColor: "primary.dark",
         borderWidth: "2px",
       },
       _disabled: {